device=device

+++ weave/neuro_functions/functions.py
@@ -707,9 +707,9 @@ def min(ten: Tensor, axis: int = None, dtype: str = None,
         use_grad: bool = False, device: str = 'cpu') -> Tensor:
     """Return the minimum of an array or minimum along an axis."""
     if ten.device == 'cpu':
-        array = np.min(a=ten.data, axis=axis, dtype=dtype)
+        array = np.min(a=ten.data, axis=axis)
     else:
-        array = cp.min(a=ten.data, axis=axis, dtype=dtype)
+        array = cp.min(a=ten.data, axis=axis)
 
     return Tensor(data=array, dtype=array.dtype, use_grad=use_grad, device=device)
